requirements.txt
python-dotenv==1.0.1
playwright==1.46.0
openpyxl==3.1.5
pandas==2.2.2
Pillow==10.3.0
requests==2.32.3
